api/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
python-socketio==5.11.3
redis==5.0.7
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.2
pydantic==2.9.2
pydantic-settings==2.5.2
argon2-cffi==23.1.0
pyotp==2.9.0
python-jose[cryptography]==3.3.0
